[Refact] #16 Auth 도메인 패키지 구성
