+++ src/main/java/com/umust/dobonglife/domain/auth/controller/AuthController.java
@@ -1,7 +1,7 @@
 package com.umust.dobonglife.domain.auth.controller;
 
-import com.umust.dobonglife.domain.auth.dto.request.RefreshTokenRequest;
-import com.umust.dobonglife.domain.auth.dto.response.TokenResponse;
+import com.umust.dobonglife.domain.auth.controller.dto.request.RefreshTokenRequest;
+import com.umust.dobonglife.domain.auth.controller.dto.response.TokenResponse;
 import com.umust.dobonglife.domain.auth.service.JwtService;
 import com.umust.dobonglife.global.common.resolver.CurrentUserId;
 import com.umust.dobonglife.global.common.response.BaseResponse;
