+++ src/main/java/com/umust/dobonglife/domain/auth/controller/dto/request/FormLoginRequest.java
@@ -1,0 +1,12 @@
+package com.umust.dobonglife.domain.auth.controller.dto.request;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class FormLoginRequest {
+    private String email;
+    private String password;
+}
