+++ src/main/java/com/umust/dobonglife/domain/auth/controller/dto/request/RefreshTokenRequest.java
@@ -1,4 +1,4 @@
-package com.umust.dobonglife.domain.auth.dto.request;
+package com.umust.dobonglife.domain.auth.controller.dto.request;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class FormLoginRequest {
-    private String email;
-    private String password;
+public class RefreshTokenRequest {
+    private String refreshToken;
 }
