--- src/main/java/com/umust/dobonglife/domain/auth/dto/request/RefreshTokenRequest.java
@@ -1,11 +1,0 @@
-package com.umust.dobonglife.domain.auth.dto.request;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RefreshTokenRequest {
-    private String refreshToken;
-}
