+++ src/main/java/com/umust/dobonglife/domain/user/domain/entity/User.java
@@ -1,6 +1,6 @@
 package com.umust.dobonglife.domain.user.domain.entity;
 
-import com.umust.dobonglife.domain.auth.model.Provider;
+import com.umust.dobonglife.domain.auth.domain.constant.Provider;
 import com.umust.dobonglife.domain.user.domain.constant.Role;
 import com.umust.dobonglife.global.common.model.BaseEntity;
 import jakarta.persistence.*;
