+++ src/main/java/com/umust/dobonglife/domain/user/domain/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.umust.dobonglife.domain.user.domain.repository;
 
-import com.umust.dobonglife.domain.auth.model.Provider;
+import com.umust.dobonglife.domain.auth.domain.constant.Provider;
 import com.umust.dobonglife.domain.user.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
